Add AcaTable, update tables returned by gen_cube
The AcaTable class is a wrapper for the astropy table, but in the future
will grow to contain a lot more specifications and also can be exported
to more formats.

The gen_cube now returns the resume table and a dictionary with each
component and meta data of the component.


Former-commit-id: ef9fd3efe42168658d7eaff43673ca8ab4d68ef4